Simple rename of ui_si5... to osc_si5... No functional change, but new name better reflects what file belongs to.
