+++ mchf-eclipse/files.mak
@@ -17,8 +17,8 @@ drivers/usb/app/usbd_audio_if.c \
 drivers/usb/app/usbd_cdc_if.c \
 drivers/usb/app/usbd_desc.c \
 drivers/ui/oscillator/soft_tcxo.c \
-drivers/ui/oscillator/ui_si570.c \
-drivers/ui/oscillator/ui_si5351a.c \
+drivers/ui/oscillator/osc_si570.c \
+drivers/ui/oscillator/osc_si5351a.c \
 drivers/ui/oscillator/osc_interface.c \
 drivers/ui/menu/ui_menu.c \
 drivers/ui/menu/ui_menu_internal.c \
